Online Pharmacy Version 4.0
Database Management Systems (COP5725) 
Project Deliverable for Phase 2:
Online Pharmacy Management System (OPMS) is a web based application which aims to provide online pharmacy services for
customers so they can easily access medicines, for pharmacists so they can manage sales and inventory with much more
efficiency and ease and for doctors so they can fill prescriptions online. This is just the basic idea of OPMS. OPMS has to offer
much more by dealing with data in the backend and provide special services for system administrators which were discussed
in the phase 1 report. This report discusses the conceptual database design in the form on an entity-relationship diagram
(ERD) and the user interface design of OPMS. The ERD gives the conceptual interaction between various entities in the system,
its relationships (with cardinalities) with each other and the attributes which, in addition to consolidating the understanding
of OPMS, will also help in the implementation of t